disable test

+++ tests/properties.hs
@@ -16,7 +16,7 @@ instance (Num a, Random a) => Arbitrary (Box V2 a) where
         y1 <- choose (0, k)
         return (Box (V2 x0 y0) (V2 x1 y1))
 
-newtype Reference = Reference { getReference :: V2 Float } deriving Show
+newtype Reference = Reference { getReference :: V2 Double } deriving Show
 
 instance Arbitrary Reference where
     arbitrary = fmap Reference $ V2 <$> oneof [pure 0, pure 0.5, pure 1] <*> oneof [pure 0, pure 0.5, pure 1]
@@ -25,14 +25,14 @@ instance Arbitrary a => Arbitrary (V2 a) where
     arbitrary = V2 <$> arbitrary <*> arbitrary
 
 prop_resize bb (getReference -> ref) sz = nearZero $ bb ^. position ref - bb' ^. position ref where
-    bb' = bb & size ref .~ getPositive sz :: Box V2 Float
+    bb' = bb & size ref .~ getPositive sz :: Box V2 Double
 
 prop_rearrange (getReference -> ref) bb pos = norm (bb ^. size 0 - bb' ^. size 0) < 10e-4 where
-    bb' = bb & position ref .~ pos :: Box V2 Float
+    bb' = bb & position ref .~ pos :: Box V2 Double
 
 prop_construct (getReference -> ref) pos (getPositive -> sz) = norm (bb ^. size 0 - sz) < 10e-4
     .&&. norm (bb ^. position ref - pos) < 10e-4 where
     
-    bb = sizePos ref # (sz, pos) :: Box V2 Float
+    bb = sizePos ref # (sz, pos) :: Box V2 Double
 
 main = $(defaultMainGenerator)
